tight_cli/blueprints/providers/aws/lambda_app/starter/requirements.txt
# tight-cli
PyYAML
pytest
boto3
botocore
marshmallow
apispec